feat: DP-6716 Update OPDP datums (#158)
[DP-6716](https://outreach-io.atlassian.net/browse/DP-6716)

Updates the `opdpManifest.schema.json` in this repo to include the
latest account-level datums.

This PR _does not_ update this service's OPDP manifest file to the very
latest version.  Some of the diffs with upstream, ie. the ones that
exclude all references to the "internalURL", look to be intentional.  I
have kept these diffs for now.